requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
idna==3.4
paho-mqtt==1.5.1
Pillow==9.4.0
psutil==5.8.0
pycodestyle==2.10.0
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
